pyproject.toml
[tool.poetry]
name = "mocking"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^23.11.0"
pylint = "^3.0.2"
pandas = "^2.1.3"
pyyaml = "^6.0.1"
icecream = "^2.1.3"

[tool.black]
line-length=79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
